backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic-settings==2.8.1
httpx==0.28.1
apscheduler==3.10.4
pandas==2.2.3
numpy==2.2.3
yfinance>=1.2.0
python-dateutil==2.9.0.post0
orjson==3.10.15
tenacity==9.0.0
pytest==8.3.4
claude-agent-sdk>=0.1.36
mcp>=1.0.0
